tests/wpt/web-platform-tests/clipboard-apis/async-write-text-read-blobtext-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Async Clipboard writeText -> read ([text/plain Blob]) tests</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async function readWriteTest(textInput) {
  promise_test(async t => {
    await navigator.clipboard.writeText(textInput);
    const blobsOutput = await navigator.clipboard.read();
    assert_equals(Object.keys(blobsOutput).length, 1);
    const blobOutput = blobsOutput['text/plain'];
    assert_equals(blobOutput.type, 'text/plain');

    const textOutput = await (new Response(blobOutput)).text();
    assert_equals(textOutput, textInput);
  }, 'Verify write and read clipboard given text: ' + textInput);
}

readWriteTest('Clipboard write text -> read ([text/plain Blob]) test');
readWriteTest('non-Latin1 text encoding test データ');
</script>
<p>
  Note: This is a manual test because it writes/reads to the shared system
  clipboard and thus cannot be run async with other tests that might interact
  with the clipboard.
</p>
